server/requirement.txt
colorama
PTable
readline
pyreadline
backports.pbkdf2
